Copyright headers plus fixed l-bfgs test
